+++ skore/src/skore/utils/_testing.py
@@ -1,0 +1,18 @@
+import contextlib
+import copy
+
+
+@contextlib.contextmanager
+def check_cache_changed(value):
+    """Assert that `value` has changed during context execution."""
+    initial_value = copy.copy(value)
+    yield
+    assert value != initial_value
+
+
+@contextlib.contextmanager
+def check_cache_unchanged(value):
+    """Assert that `value` has not changed during context execution."""
+    initial_value = copy.copy(value)
+    yield
+    assert value == initial_value
