+++ skore/tests/unit/sklearn/estimator/feature_importance/test_permutation_importance.py
@@ -1,4 +1,3 @@
-import contextlib
 import copy
 
 import numpy as np
@@ -12,22 +11,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.pipeline import make_pipeline
 from sklearn.preprocessing import StandardScaler
 from skore import EstimatorReport
-
-
-@contextlib.contextmanager
-def check_cache_changed(value):
-    """Assert that `value` has changed during context execution."""
-    initial_value = copy.copy(value)
-    yield
-    assert value != initial_value
-
-
-@contextlib.contextmanager
-def check_cache_unchanged(value):
-    """Assert that `value` has not changed during context execution."""
-    initial_value = copy.copy(value)
-    yield
-    assert value == initial_value
+from skore.utils._testing import check_cache_changed, check_cache_unchanged
 
 
 def regression_data():
